requirements.txt
numpy
pandas
joblib
pyyaml
tqdm
matplotlib
tables
seaborn